feat: add benchmark_vs_grep.sh script for performance comparison
Add a new script, benchmark_vs_grep.sh, that performs a performance comparison between the 'grep' command and our 'go-grep' command. The script installs the latest version of 'go-grep', retrieves the current commit hash, and runs 'grep' and 'go-grep' commands on dummy data. The results are logged in the benchmark_vs_grep_stats.log file.

This script will help us evaluate the performance of our 'go-grep' command and compare it with the standard 'grep' command as the project progresses.